added enemy spwaning first try at animation of enemy spwaning

+++ Assets/Entities/Player/Projectile.cs
@@ -3,8 +3,7 @@ using System.Collections;
 
 public class Projectile : MonoBehaviour {
 
-
-	public float Damage;
+	public int Damage;
 
 	public void Hit()
 	{
